Fully Functional Sample App
The app retrieves a json list using Retrofit and RxJava for callbacks and manipulating the retrieved data, it is then converted to LiveData to display on the UI. Architecture pattern used is MVVM. Viewmodels are instantiated through the use of the ViewModelProviders class to link them to the lifecycle of their acitivities. All other classes are injected via Dagger2. A ViewModelFactory class was created and injected via Dagger to allow the functioning of ViewModelProviders along with Dagger.